update  notifications
remove LevelTypes.CUSTOM to clearify the distinction between a level-based
notification and a fixed notification. A fixed notification is always a
sort of custom notification (content is string or TemplateResult),
permanent (without autoclose-time, triggered by user-interaction)
and should stay in a position where the current tasks and
actions of the user are not interrupted by the notification, but are
recognizable. Fixed notification are made for optional tasks and
decisions, which are not  essential for the current workflow, so a
cancelation with any user-interaction is enough as criteria.
A level-based notification is made for informations which are
needed to presented to the user without any need of a following
user-decision.

add/update tests for notifiactions.reducer

add debounce behavior in NotificationPlugIn for clear a fixed
notification

update notification-example in ShowCase