update behavior on text-changes
update to the lifecycle of the text-style. This includes the
initialization of the styleOptions with a default-text (i18n),
checking for initial states (styleOption.text==null) and
prevent empty strings in the input by reset the text-property to null.

olStyleUtils.js: update usage of DEFAULT_TEXT to cases where
styleOption.text is not a String, Empty string are allowed.

OlDrawHandler.js: update _getStyleOption, to guarantee a valid
styleOption.text

DrawToolContent.js: fix the order of activate-steps for a drawTool;
add sanity-check to prevent text-features with empty text