refactor: extract global constant , add id-test
The constant for a temporary  GeoResourceId (temp_measure_id)
for temporary drawings and measurements in offline -mode becomes a wider
range of usage. Due to the fact, that this GeoResourceId is equivalent
to a FileStorageId, the extraction and renaming takes place
(temp_storage_id).
Creating additionally id-check, whether a GeoResourceId is a StorageId
in general or not, including the new global constant.

Add tests