extract impl of getPadding to paddingProvider
extract implementation of getPadding to paddingProvider, to get the
flexibility for future changes or improvements.
The currently used  implementation 'mainMenuLandscapePaddingProvider'
relies on states in the store and style-properties which would taint the
MapService with the implicit dependencies of StoreService and the custom
style property variable ('--width-mainmenu)'.

add tests for padding.provider and MapService.getPadding()